pyproject.toml
[tool.poetry]
name = "AI Code Documentation Generator"
version = "0.1.0"
description = "AI-powered Code Documentation Generator"
authors = ["Jim Gibbs <jim@nedivo.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
trio = "^0.23.2"
requests = "^2.31.0"
openai = "^1.3.9"
nltk = "^3.8.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
flake8 = "^6.1.0"
pre-commit = "^3.6.0"
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
